move step manager to debugger folder
